+++ lisp/eab-words.el
@@ -19,6 +19,9 @@
 (eab/search-word eab/edit-list-at-point "Edit-list at point" eab/edit-list-1 (word))
 (eab/search-word eab/gnugol-search "gnugol search" gnugol (word))
 (eab/search-word eab/org-agenda-search "agenda search" org-search-view (nil word))
+;; (eab/search-word eab/org-agenda-search
+;; 		 "agenda ql search" org-ql-search
+;; 		 (org-agenda-files '(and (clocked 560) (not (tags "noagenda"))) :super-groups '((:auto-dir-name)) :sort 'priority))
 (eab/search-word eab/twit-search "twitter search" twittering-search (word)) ;; SOMEDAY команда сохраняется в command-history и её можно повторить
 (eab/search-word eab/add-eab-abbrev "abbrev save" define-abbrev (eab-abbrev-table word word))
 (eab/search-word eab/google-translate "google translate" eab/shell-translate (word))